+++ cli/src/templates/project/client-sdk/index.ts
@@ -1,1 +1,1 @@
-export * from "./universalSdk";
+export * from "./clientSdk";