+++ blueprint/implementations/worker/database/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/implementation-worker-database
 
+## 0.8.12
+
+### Patch Changes
+
+- minor typing updates
+- Updated dependencies
+  - @forklaunch/interfaces-worker@0.7.12
+
 ## 0.8.11
 
 ### Patch Changes